\name{AQstats}
\alias{AQstats}

\title{
Generate a SpatialPoints object corresponding to the best result obtained in an optimized network 
}
\description{
Generate a 
}
\usage{
AQstats(data)
}

\arguments{
  \item{data}{
object of class rbga resulting from 
}
}
\value{
a SpatialPoints object
}
\seealso{
See function 
}
\examples{
data(simData)
resAQ <- AQstats(simData)
resAQ
}

\keyword{ spatial }
