\name{BuildHeader}
\alias{BuildHeader}
\title{BuildHeader}
\usage{
BuildHeader()
}
\value{
Header string for the old API auth method.
}
\description{
Internal function - Build Header for REST API call. This is
redundant, we should be using OAUTH.
}
\seealso{
Other internal: \code{\link{ApiGetElements}};
\code{\link{ApiGetEvars}}; \code{\link{ApiGetMetrics}};
\code{\link{ApiGetProps}};
\code{\link{ApiGetReportSuites}};
\code{\link{ApiGetReport}}; \code{\link{ApiGetSegments}};
\code{\link{ApiGetSuccessEvents}};
\code{\link{ApiQueueReport}};
\code{\link{ApiValidateReport}};
\code{\link{BuildInnerBreakdownsRecursively}};
\code{\link{BuildURL}}; \code{\link{ParseFallout}};
\code{\link{ParseOvertime}}; \code{\link{ParsePathing}};
\code{\link{ParseRanked}}; \code{\link{ParseTrended}};
\code{\link{PostRequest}}
}

