\name{FigQAmultiRes}
\alias{FigQAmultiRes}

\title{
Generate a SpatialPoints object corresponding to the best result obtained in an optimized network 
}
\description{
Generate a 
}
\usage{
FigQAmultiRes(x, y, mask, fact=2)
}

\arguments{
  \item{x}{
object of class rbga resulting from 
}
  \item{y}{
object of class rbga resulting from 
}
  \item{mask}{
object of class rbga resulting from 
}
  \item{fact}{
object of class rbga resulting from 
}
}
\value{
a SpatialPoints object
}
\seealso{
See function 
}
\examples{
#data(example_x1)
#data(example_y1)
#data(example_mask1)
#FigQAmultiRes(example_x1, example_y1, example_mask1, fact=2) 
}

\keyword{ spatial }
