\name{GetSuccessEvents}
\alias{GetSuccessEvents}
\title{GetSuccessEvents}
\usage{
GetSuccessEvents(reportsuite.ids)
}
\arguments{
  \item{reportsuite.ids}{report suite id (or list of report
  suite ids)}
}
\value{
List of valid successevents
}
\description{
Gets available report suite successevents
}

