\name{MAE}
\alias{MAE}

\title{
Generate a SpatialPoints object corresponding to the best result obtained in an optimized network 
}
\description{
Generate a 
}
\usage{
MAE(layer1, layer2)
}

\arguments{
  \item{layer1}{
object of class rbga resulting from 
}
  \item{layer2}{
object of class rbga resulting from 
}
}
\value{
a SpatialPoints object
}
\seealso{
See function 
}

\keyword{ spatial }
