\name{ParsePathing}
\alias{ParsePathing}
\title{ParsePathing}
\usage{
ParsePathing(report.data)
}
\arguments{
  \item{report.data}{jsonlite formatted data frame of
  report data returned from the API}
}
\value{
Formatted data frame
}
\description{
Internal Function - Parses a pathing report returned from
the API
}

