\name{ParseRanked}
\alias{ParseRanked}
\title{ParseRanked}
\usage{
ParseRanked(report.data)
}
\arguments{
  \item{report.data}{jsonlite formatted data frame of
  report data returned from the API}
}
\value{
Formatted data frame
}
\description{
Internal Function - Parses a ranked report returned from
the API
}

