\name{quantityDisagree}
\alias{quantityDisagree}

\title{
Generate a SpatialPoints object corresponding to the best result obtained in an optimized network 
}
\description{
Generate a 
}
\usage{
quantityDisagree(data)
}

\arguments{
  \item{data}{
object of class rbga resulting from 
}
}
\value{
a SpatialPoints object
}
\seealso{
See function 
}

\keyword{ spatial }
