\name{rAA-package}
\alias{rAA-package}
\alias{rAA}
\docType{package}
\title{
Accuracy assessment
}
\description{
This package contains functions for assessment of accuracy}
\details{
\tabular{ll}{
Package: \tab rAA\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2012-12-31\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Robert G. Pontius <rpontius@clarku.edu>, Alí Santacruz

Maintainer: Alí Santacruz <asantacruzdelgado@clarku.edu>
}
\keyword{ package }
\keyword{ spatial }
\seealso{
\code{\link{AQstats}}, \code{\link{AIQImat}} 
}
