\docType{package}
\name{RAA}
\alias{RAA}
\alias{RAA-package}
\title{R Client for Adobe Analytics v1.4 API}
\description{
R Client for Adobe Analytics v1.4 API
}
\details{
\tabular{ll}{ Package: \tab RAA\cr Type: \tab Package\cr
Version: \tab 0.1\cr Date: \tab 2014-02-03\cr License: \tab
GPL (>= 2)\cr LazyLoad: \tab yes\cr }

R Client for Adobe Analytics v1.4 Reporting API. Partly
based on RSiteCatalyst by Randy Zwitch.
}
\author{
Willem Paling \email{willem.paling
}
\keyword{package}

