\name{simData}
\alias{simData}
\docType{data}
\title{
Map of total soil carbon stock (t/ha) at 0-10 cm depth
}
\description{
Map of total soil 
}
\usage{data(simData)}
\format{
  The format is:
Formal class 'SpatialPixelsDataFrame' [package "sp"] 
}
\source{
Santacruz, A., 2010. 
}
\references{
Santacruz, A., 2011. \ 
}
\seealso{

}
\examples{
data(simData)
resAQ <- AQstats(simData)
resAQ
}
\keyword{datasets}
