\name{zonalMap}
\alias{zonalMap}

\title{
Generate a SpatialPoints object corresponding to the best result obtained in an optimized network 
}
\description{
Generate a 
}
\usage{
zonalMap(layer, features)
}

\arguments{
  \item{layer}{
object of class rbga resulting from 
}
  \item{features}{
object of class rbga resulting from 
}
}
\value{
a SpatialPoints object
}
\seealso{
See function 
}

\keyword{ spatial }
